Move ToolChain to different module
This allows early-bird tools like 'install' and 'path' to run without clint installed
